api_auto_v2/requirements.txt
# pytest相关
pytest==5.3.2

# allure相关
allure_python==2.8.6
allure_python_commons==2.8.6

# MySQL相关
PyMySQL==0.9.3
SQLAlchemy==1.3.4
sqlacodegen==2.1.0

# Redis相关
redis==3.3.11

# Requests相关
requests==2.22.0

# flask相关
Flask==1.0.2
flask_restplus==0.13.0

# 环境读取
python-dotenv==0.10.3
